123 arriba

+++ src/VoucherWarehouse.Application/Modules/VoucherWarehouse/EcfVoucherWarehouse/Dto/EcfVoucherWarehouseCreateDto.cs
@@ -2,6 +2,6 @@
 
 namespace IBS.VoucherWarehouse.Modules.VoucherWarehouse.EcfVoucherWarehouse.Dto;
 
-public sealed record class EcfVoucherWarehouseCreateDto : BaseEntityDto<int>
+public sealed record class EcfVoucherWarehouseCreateDto : BaseEntityDto<long>
 {
 }
